+++ src/itype.jl
@@ -3,7 +3,7 @@
 # returns the type of the atom, that is, the index of this atom within the molecule
 # (goes from 1 to natomspermol)
 #
-function itype(iatom :: Int64, natomspermol :: Int64)
+function itype(iatom :: Int, natomspermol :: Int)
   itype = iatom%natomspermol
   if itype == 0
     itype = natomspermol
@@ -12,4 +12,4 @@ function itype(iatom :: Int64, natomspermol :: Int64)
 end
 
 # Calling using the structures of Solute and Solvent, to clear up the code above
-itype(iatom :: Int64, s :: Selection) = itype(iatom,s.natomspermol)
+itype(iatom :: Int, s :: Selection) = itype(iatom,s.natomspermol)
